REFACTOR: split reactapp blueprint into multiple bps dep on function
