+++ app/factory.py
@@ -6,7 +6,9 @@ from flask import Flask
 from flask_bootstrap import Bootstrap
 import config
 from routes.views import bp as spfy
-from routes.reactapp import bp_ra as reactapp
+from routes.ra_views import bp_ra_views
+from routes.ra_posts import bp_ra_posts
+from routes.ra_statuses import bp_ra_statuses
 from flask_recaptcha import ReCaptcha
 from flask_cors import CORS, cross_origin
 
@@ -27,6 +29,9 @@ def create_app():
     cors = CORS(app, resources={r"/api/*": {"origins": "*"}})
 
     app.register_blueprint(spfy)
-    app.register_blueprint(reactapp)
+    # register the new blueprints used by reactapp
+    app.register_blueprint(bp_ra_views)
+    app.register_blueprint(bp_ra_posts)
+    app.register_blueprint(bp_ra_statuses)
 
     return app
