+++ app/routes/ra_posts.py
@@ -1,0 +1,132 @@
+import os
+import tarfile
+import zipfile
+import redis
+from datetime import datetime
+from flask import Blueprint, request, jsonify, current_app
+from flask_recaptcha import ReCaptcha
+from werkzeug.utils import secure_filename
+from routes.file_utils import fix_uri, handle_tar, handle_zip
+from modules.gc import blob_gc_enqueue
+from modules.spfy import spfy
+
+bp_ra_posts = Blueprint('reactapp_posts', __name__)
+
+def handle_groupresults(jobs_dict):
+    '''
+    if we're grouping results for the Front-End
+    take all jobs in jobs_dict and store it in Redis (not RQ)
+    respond with a key to fetch this jobs_dict from Redis
+    relies on a new check in /results/<job_id> which is able to poll RQ
+    tasks on behalf of the front-end
+        Args:
+            jobs_dict (dict): a dictionary where job ids are the keys and
+                the values are also dictionaries with keys 'analysis' and 'file'
+        Return:
+            (dict): a dictionary with key as novel jobid of form 'blob' + hash
+                meant to be parsable by same code as for old, non-grouped ver
+    '''
+    # generate a novel job id
+    # we prepend 'blob' so the /results path can tell its our custom object
+    # and shouldnt be retrieved from RQ
+    job_id = 'blob' + str(hash(str(jobs_dict)))
+    # start a redis connection
+    redis_url = current_app.config['REDIS_URL']
+    redis_connection = redis.from_url(redis_url)
+    #with redis.from_url(redis_url) as redis_connection:
+    # set the job_id: jobs_dict pair in Redis
+    redis_connection.set(job_id, jobs_dict)
+    # create a similar structure to the old return
+    d = {}
+    d[job_id] = {}
+    d[job_id]['analysis'] = "Subtyping"
+    st = set()
+    for key in jobs_dict:
+        st.add(jobs_dict[key]['file'])
+    s = ''
+    for f in st:
+        s += f + ' '
+    d[job_id]['file'] = s
+    return d
+
+# the /api/v0 prefix is set to allow CORS for any postfix
+# this is a modification of the old upload() methods in views.py
+@bp_ra_posts.route('/api/v0/upload', methods=['POST'])
+def upload():
+    recaptcha = ReCaptcha(app=current_app)
+    if recaptcha.verify():
+        form = request.form
+        options = {}
+        # defaults
+        options['amr']=True
+        options['vf']=True
+        options['serotype']=True
+        options['pi']=90
+        # new to 4.2.0
+        # we consider False as default as the front-end should override this
+        # to use the new feature
+        groupresults = False
+
+        # processing form data
+        for key, value in form.items():
+            #we need to convert lower-case true/false in js to upper case in python
+                #remember, we also have numbers
+            if not value.isdigit():
+                if value.lower() == 'false':
+                    value = False
+                else:
+                    value = True
+                if key == 'options.amr':
+                    options['amr']=value
+                if key == 'options.vf':
+                    options['vf']=value
+                if key == 'options.serotype':
+                    options['serotype']=value
+                if key == 'options.groupresults':
+                    groupresults = value
+            else:
+                if key =='options.pi':
+                    options['pi']=int(value)
+
+        # get a list of files submitted
+        uploaded_files = request.files.getlist("file")
+        print uploaded_files
+
+        #set up constants for identifying this sessions
+        now = datetime.now()
+        now = now.strftime("%Y-%m-%d-%H-%M-%S-%f")
+        jobs_dict = {}
+
+        for file in uploaded_files:
+            if file:
+                # for saving file
+                filename = os.path.join(current_app.config[
+                                        'DATASTORE'], now + '-' + secure_filename(file.filename))
+                file.save(filename)
+                print 'Uploaded File Saved at', str(filename)
+
+                if tarfile.is_tarfile(filename):
+                    # set filename to dir for spfy call
+                    filename = handle_tar(filename, now)
+                elif zipfile.is_zipfile(filename):
+                    filename = handle_zip(filename, now)
+
+                # for enqueing task
+                jobs_enqueued = spfy(
+                    {'i': filename, 'disable_serotype': False, 'disable_amr': False, 'disable_vf': False, 'pi':options['pi'], 'options':options})
+                jobs_dict.update(jobs_enqueued)
+        # new in 4.2.0
+        if groupresults:
+            return jsonify(handle_groupresults(jobs_dict))
+        else:
+            return jsonify(jobs_dict)
+    else:
+        return "Captcha Failed Verification", 500
+
+# this is for the Group Comparisons (Fishers) module
+@bp_ra_posts.route('/api/v0/newgroupcomparison', methods=['POST'])
+def handle_group_comparison_submission():
+    query = request.json['groups']
+    target = request.json['target']
+    jobid = blob_gc_enqueue(query, target)
+    return jobid
